requirements.txt
asgiref==3.9.1
Django==5.2.4
gunicorn==23.0.0
packaging==25.0
Pillow==10.0.0
sqlparse==0.5.3
